% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fetchTME}
\alias{fetchTME}
\title{Fetches all components of a TME dataset}
\usage{
fetchTME(df, row, sparse)
}
\arguments{
\item{df}{the metadata dataframe}

\item{row}{the row storing the TME dataset}

\item{sparse}{whether to download as a sparse matrix or not}
}
\description{
Fetches all components of a TME dataset
}
