% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryTME.R
\name{queryTME}
\alias{queryTME}
\title{A function to query TME datasets available in this package}
\usage{
queryTME(
  geo_accession = NULL,
  score_type = NULL,
  has_signatures = NULL,
  has_truth = NULL,
  tumour_type = NULL,
  metadata_only = FALSE
)
}
\arguments{
\item{geo_accession}{Search by geo accession number}

\item{metadata_only}{Return rows of metadata instead of actual datasets. Defaults to FALSE}
}
\description{
This function allows you to search and subset included TME datasets
}
\examples{
queryTME
}
\keyword{tumour}
