% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryTME.R
\name{queryTME}
\alias{queryTME}
\title{A function to query TME datasets available in this package}
\usage{
queryTME(
  geo_accession = NULL,
  score_type = NULL,
  has_signatures = NULL,
  has_truth = NULL,
  tumour_type = NULL,
  author = NULL,
  journal = NULL,
  year = NULL,
  pmid = NULL,
  sequence_tech = NULL,
  organism = NULL,
  metadata_only = FALSE,
  sparse = FALSE
)
}
\arguments{
\item{geo_accession}{Search by geo accession number}

\item{score_type}{Search by type of score (TPM, FPKM, raw count)}

\item{has_signatures}{Return datasets that have gene signatures available (TRUE/FALSE)}

\item{has_truth}{Search by the presence of known cell-type labels}

\item{tumour_type}{Search by type of tumour contained in the dataset}

\item{author}{Search by author}

\item{journal}{Search by journal}

\item{year}{Search by exact year or year ranges with '<', '>', or '-'}

\item{pmid}{Search by Pubmed ID}

\item{sequence_tech}{Search by sequencing technology}

\item{organism}{Search by source organism}

\item{metadata_only}{Return rows of metadata instead of actual datasets. Defaults to FALSE}

\item{sparse}{Return expression as a sparse matrix.
Uses less memory but is less convenient to view. Defaults to FALSE.}
}
\value{
A list containing a table of metadata or
one or more SingleCellExperiment objects
}
\description{
This function allows you to search and subset included TME datasets
}
\examples{
res <- queryTME(metadata_only = TRUE)
\dontrun{
res <- queryTME(geo_accession = "GSE72056")
res <- queryTME(has_truth = TRUE, has_signatures = TRUE)
}
}
\keyword{tumour}
