% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTME.R
\name{saveTME}
\alias{saveTME}
\title{A function to save a TME dataset}
\usage{
saveTME(object, outdir)
}
\arguments{
\item{object}{The tme_data object to be written to disk}

\item{outdir}{The directory to save the tme_data in}
}
\value{
Nothing
}
\description{
This function allows you to save the expression,
labels, and cell types to disk in csv format
}
\examples{
\dontrun{res <- queryTME(geo_accession = 'GSE72056')[[1]]}
\dontshow{res <- SingleCellExperiment(list(counts=matrix()))
tdir = tempdir()
filename = file.path(tdir, 'save_tme_data')} 
saveTME(res, filename)
}
\keyword{tumour}
